Make the compaction log fetching smarter and less memory hungry
Instead of fetching the complete region server log file (which can be
in the hundreds of MBs), we initially look back only 1024 KBs prior to
the end of the file. Each subsequent refresh of the compaction metrics
starts at the offset of the last new line of the previous fetch. This
makes each fetch of the compaction logs faster and lowers the memory
requirement to keep the content in memory for parsing.