Add Defaulting of hbase Configuration to the System Config.
Update the create_package script so that the hannibal startup
script gets the hbase configuration from the system it is
running from using 'hbase classpath'. This allows hannibal to be
installed without manual hbase-site.xml copying/editing
on a gateway node that has the target hbase cluster configuration.

In a case where the 'hbase shell' command is not available,
we get the same behavior as before. Also note that a configuration
file present in conf/hbase-site.xml should override the hbase
system configuration since it's before in the classpath.